requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
blosc2==2.7.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
colorlog==6.8.2
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
executing==2.0.1
Farama-Notifications==0.0.4
fastjsonschema==2.20.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
fsspec==2024.9.0
furl==2.1.3
gast==0.6.0
google-api-core==2.21.0
google-auth==2.35.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.67.0
gym-notices==0.0.8
gymnasium==0.28.1
h11==0.14.0
h2o==3.46.0.5
h5py==3.11.0
httpcore==1.0.6
httpx==0.27.2
idna==3.7
imbalanced-learn==0.12.4
ipykernel==6.29.5
ipython==8.26.0
isoduration==20.11.0
itsdangerous==2.2.0
jax-jumpy==1.0.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.4.1
keras-metrics==1.1.0
kiwisolver==1.4.7
libclang==18.1.1
libtaxii==1.1.119
lxml==5.3.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.0.8
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
ndindex==1.8
nest-asyncio==1.6.0
networkx==3.4.1
notebook==7.2.2
notebook_shim==0.2.4
numexpr==2.10.1
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
orderedmultidict==1.0.1
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
plotly==5.24.1
prometheus_client==0.21.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==70.2.0
Shimmy==1.0.0
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
stix2-patterns==2.0.0
sympy==1.13.1
tables==3.9.2
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
torch==2.5.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.2
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
